pyproject.toml
[tool.poetry]
name = "pizza"
version = "0.1.0"
description = ""
authors = ["Ali-hassan23 <127784171+Ali-hassan23@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
uvicorn = "^0.30.6"
fastapi = "^0.112.0"
sqlalchemy = "^2.0.32"
psycopg2 = "^2.9.9"
pydantic = "^2.8.2"
sqlalchemy-utils = "^0.41.2"
python-multipart = "^0.0.9"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
jwt = "^1.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
